backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.1
bcrypt==4.1.1
python-multipart==0.0.6
ultralytics==8.0.234
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
cors==1.0.1
sqlalchemy==2.0.23
psycopg==3.14.0
asyncpg==0.29.0
httpx==0.25.2
